requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
Django==2.2
djangorestframework==3.10.3
dnspython==2.2.1
idna==3.4
pymongo==3.13.0
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
